rename: IntegrationTest 어노테이션 패키지 위치 변경
